fix: apiFunctions.test.tsx
