+++ api/apiFunctions.test.tsx
@@ -1,11 +1,10 @@
-import { log } from 'console';
 import {
   addUserToLeague,
   getAllLeagues,
-  getUserDocumentId,
   loginAccount,
   logoutAccount,
   registerAccount,
+  getCurrentLeague,
 } from './apiFunctions';
 import { account } from './config';
 const apiFunctions = require('./apiFunctions');
@@ -21,15 +20,13 @@ jest.mock('./apiFunctions', () => {
     getAllLeagues: jest.fn(),
     getUserDocumentId: jest.fn(),
     addUserToLeague: jest.fn(),
+    getCurrentLeague: jest.fn(),
+    loginAccount: jest.fn(),
+    registerAccount: jest.fn(),
   };
 });
 
 describe('Auth Functions', () => {
-  jest.mock('./apiFunctions', () => ({
-    loginAccount: jest.fn(),
-    registerAccount: jest.fn(),
-  }));
-
   xdescribe('login account successful', () => {
     it('should show user login successfully', async () => {
       const userDummy = {
@@ -63,7 +60,7 @@ describe('Auth Functions', () => {
   });
 
   account.create = jest.fn();
-  describe('register account successful', () => {
+  xdescribe('register account successful', () => {
     it('Should allow a user to register an account', async () => {
       const userDummy = {
         email: 'testemail0@email.com',
